各モジュール実装を処理を含まないtypeで書き換え
