--- src/simulations/modular_life_v2/module/any_module.ts
@@ -1,5 +1,0 @@
-import { Assemble } from "./assemble"
-import { Compute } from "./compute"
-import { Hull } from "./hull"
-
-export type AnyModule = Assemble | Compute | Hull
