--- src/simulations/modular_life_v2/module/index.ts
@@ -1,5 +1,0 @@
-export * from "./any_module"
-export * from "./assemble"
-export * from "./compute"
-export * from "./hull"
-export * from "./types"
