+++ src/simulations/modular_life_v2/system.ts
@@ -1,3 +1,3 @@
 export const System = {
-  version: "2.0.3"
+  version: "2.0.4"
 } as const
